Move regression and classification trees into packages
