+++ src/test/scala/io/citrine/lolo/trees/classification/ClassificationTreeTest.scala
@@ -1,11 +1,11 @@
-package io.citrine.lolo.trees
+package io.citrine.lolo.trees.classification
 
 import java.io.{File, FileOutputStream, ObjectOutputStream}
 
 import io.citrine.lolo.TestUtils
 import io.citrine.lolo.stats.functions.Friedman
-import org.scalatest.Assertions._
 import org.junit.Test
+import org.scalatest.Assertions._
 
 import scala.util.Random
